requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
future==0.18.2
google-auth==2.6.6
google-auth-oauthlib==0.5.1
greenlet==1.1.2
gspread==5.3.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.2.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0